Move string dtype conversions to h5dataset
Moving the string conversion down into the H5DataSet class is more
appropriate:
- The data type conversion is HDF5-specific: We convert the dtype of
  string objects from disk into Python strings to be returned to the
  upper layers of the library.  Other (hypothetical, future)
  implementations would likely have other requirements.
- It avoids needing to make the conversion in all the places where data
  is read.
